Lots of stuff.
Right comic will now be loaded when the player collides with an enemy or an entrance. Based on tags and object names.
Need the comics and the prefabs setup before this does much of anything though.
Spawn location is now stored in the GameManager for spawning the player in the right location after comics.
And more...